requirements.txt
langdetect
mysql-connector
